Add remote container/image filter and removal
