+++ core/utils/cleanup.go
@@ -3,9 +3,12 @@ package utils
 import (
 	"fmt"
 
+	container_types "github.com/docker/docker/api/types"
+	"github.com/sdslabs/beastv4/core/config"
 	cfg "github.com/sdslabs/beastv4/core/config"
 	"github.com/sdslabs/beastv4/core/database"
 	"github.com/sdslabs/beastv4/pkg/cr"
+	"github.com/sdslabs/beastv4/pkg/remoteManager"
 
 	log "github.com/sirupsen/logrus"
 )
@@ -14,17 +17,29 @@ func CleanupContainerByFilter(filter, filterVal string) error {
 	if filter != "id" && filter != "name" {
 		return fmt.Errorf("Not a valid filter %s", filter)
 	}
-
-	containers, err := cr.SearchContainerByFilter(map[string]string{filter: filterVal})
+	var containers []container_types.Container
+	var err error
+	containers, err = cr.SearchContainerByFilter(map[string]string{filter: filterVal})
 	if err != nil {
 		log.Error("Error while searching for container with %s : ", filter, filterVal)
 		return err
 	}
+	remoteContainers, err := remoteManager.SearchContainerByFilterRemote(map[string]string{filter: filterVal})
+	if err != nil {
+		log.Error("Error while searching for remote container with %s : ", filter, filterVal)
+		return err
+	}
+	containers = append(containers, remoteContainers...)
 
 	var erroredContainers []string
 	if len(containers) != 0 {
 		log.Infof("Cleaning up container with %s %s", filter, filterVal)
 		for _, container := range containers {
+			err = remoteManager.StopAndRemoveContainerRemote(container.ID, config.AvailableServer{})
+			if err != nil {
+				erroredContainers = append(erroredContainers, container.ID)
+				log.Errorf("Error while cleaning up remote container %s : %s", container.ID, err)
+			}
 			err = cr.StopAndRemoveContainer(container.ID)
 			if err != nil {
 				erroredContainers = append(erroredContainers, container.ID)
@@ -55,10 +70,19 @@ func CleanupChallengeContainers(chall *database.Challenge, config cfg.BeastChall
 }
 
 func CleanupChallengeImage(chall *database.Challenge) error {
-	err := cr.RemoveImage(chall.ImageId)
-	if err != nil {
-		log.Error("Error while cleaning up image with id ", chall.ImageId)
-		return err
+	if chall.ServerDeployed != "localhost" {
+		server := config.Cfg.AvailableServers[chall.ServerDeployed]
+		err := remoteManager.RemoveImageRemote(chall.ImageId, server)
+		if err != nil {
+			log.Error("Error while cleaning up image on remote %s with id ", chall.ServerDeployed, chall.ImageId)
+			return err
+		}
+	} else {
+		err := cr.RemoveImage(chall.ImageId)
+		if err != nil {
+			log.Error("Error while cleaning up image with id ", chall.ImageId)
+			return err
+		}
 	}
 
 	database.UpdateChallenge(chall, map[string]interface{}{"ImageId": GetTempImageId(chall.Name)})
