+++ core/utils/logs.go
@@ -4,9 +4,12 @@ import (
 	"errors"
 	"fmt"
 
+	container_types "github.com/docker/docker/api/types"
 	"github.com/sdslabs/beastv4/core"
+	"github.com/sdslabs/beastv4/core/config"
 	"github.com/sdslabs/beastv4/core/database"
 	"github.com/sdslabs/beastv4/pkg/cr"
+	"github.com/sdslabs/beastv4/pkg/remoteManager"
 )
 
 func GetLogs(challname string, live bool) (*cr.Log, error) {
@@ -22,12 +25,16 @@ func GetLogs(challname string, live bool) (*cr.Log, error) {
 	if !IsContainerIdValid(chall.ContainerId) {
 		return nil, fmt.Errorf("Underlying challenge configuration present is not valid.")
 	}
-
-	containers, err := cr.SearchContainerByFilter(map[string]string{"id": chall.ContainerId})
+	var containers, remoteContainers []container_types.Container
+	remoteContainers, err = remoteManager.SearchContainerByFilterRemote(map[string]string{"id": chall.ContainerId})
+	if err != nil {
+		return nil, fmt.Errorf("Error while searching for remote container with id %s", chall.ContainerId)
+	}
+	containers, err = cr.SearchContainerByFilter(map[string]string{"id": chall.ContainerId})
 	if err != nil {
 		return nil, fmt.Errorf("Error while searching for container with id %s", chall.ContainerId)
 	}
-
+	containers = append(containers, remoteContainers...)
 	if len(containers) > 1 {
 		return nil, errors.New("Got more than one containers, something fishy here. Contact admin to check manually.")
 	}
@@ -37,9 +44,17 @@ func GetLogs(challname string, live bool) (*cr.Log, error) {
 	}
 
 	if live {
-		cr.ShowLiveContainerLogs(chall.ContainerId)
+		if chall.ServerDeployed != "localhost" {
+			server := config.Cfg.AvailableServers[chall.ServerDeployed]
+			remoteManager.ShowLiveContainerLogsRemote(chall.ContainerId, server)
+		} else {
+			cr.ShowLiveContainerLogs(chall.ContainerId)
+		}
 		return nil, nil
 	}
-
+	if chall.ServerDeployed != "localhost" {
+		server := config.Cfg.AvailableServers[chall.ServerDeployed]
+		return remoteManager.GetContainerStdLogsRemote(chall.ContainerId, server)
+	}
 	return cr.GetContainerStdLogs(chall.ContainerId)
 }
